+++ src/panel_save.rs
@@ -1,9 +1,11 @@
+use std::path::PathBuf;
+
 use eframe::egui;
 
 use crate::panel_export::TEXTEDIT_WIDTH;
 pub struct PanelSaveLoad {
-    pub save_dir: String,
-    pub file_name: String,
+    pub file_path: String,
+    cur_dir: PathBuf,
 }
 
 pub enum SaveLoadAction {
@@ -13,39 +15,44 @@ pub enum SaveLoadAction {
 
 impl Default for PanelSaveLoad {
     fn default() -> Self {
-        Self {
-            save_dir: std::env::current_dir().unwrap().display().to_string(),
-            file_name: "my_terrain.wgen".to_owned(),
-        }
+        let cur_dir = std::env::current_dir().unwrap();
+        let file_path = format!("{}/my_terrain.wgen", cur_dir.display().to_string());
+        Self { file_path, cur_dir }
     }
 }
 
 impl PanelSaveLoad {
-    pub fn get_file_path(&self) -> String {
-        format!("{}/{}", self.save_dir, self.file_name)
+    pub fn get_file_path(&self) -> &str {
+        &self.file_path
     }
     pub fn render(&mut self, ui: &mut egui::Ui) -> Option<SaveLoadAction> {
+        let mut action = None;
         ui.heading("Save/load project");
-        ui.label("Save/Load directory");
-        ui.add(
-            egui::TextEdit::singleline(&mut self.save_dir)
-                .hint_text("Directory")
-                .desired_width(TEXTEDIT_WIDTH),
-        );
-        ui.label("File name");
         ui.horizontal(|ui| {
-            ui.add(
-                egui::TextEdit::singleline(&mut self.file_name)
-                    .hint_text("File name")
-                    .desired_width(TEXTEDIT_WIDTH),
-            );
+            ui.label("File path");
+            if ui.button("Pick...").clicked() {
+                if let Some(path) = rfd::FileDialog::new()
+                    .set_directory(&self.cur_dir)
+                    .pick_file()
+                {
+                    self.file_path = path.display().to_string();
+                    self.cur_dir = if path.is_file() {
+                        path.parent().unwrap().to_path_buf()
+                    } else {
+                        path
+                    };
+                }
+            }
         });
-        if ui.button("Load!").clicked() {
-            return Some(SaveLoadAction::Load);
-        }
-        if ui.button("Save!").clicked() {
-            return Some(SaveLoadAction::Save);
-        }
-        None
+        ui.add(egui::TextEdit::singleline(&mut self.file_path).desired_width(TEXTEDIT_WIDTH));
+        ui.horizontal(|ui| {
+            if ui.button("Load!").clicked() {
+                action = Some(SaveLoadAction::Load);
+            }
+            if ui.button("Save!").clicked() {
+                action = Some(SaveLoadAction::Save);
+            }
+        });
+        action
     }
 }